[ISSUE-3072] Moving creation logic to StreamOfflinePluginFactory
