+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/application/ChatInitializer.kt
@@ -6,10 +6,8 @@ import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.logger.ChatLogLevel
 import io.getstream.chat.android.client.notifications.handler.NotificationConfig
 import io.getstream.chat.android.client.notifications.handler.NotificationHandlerFactory
-import io.getstream.chat.android.client.setup.InitializationCoordinator
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
-import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.markdown.MarkdownTextTransformer
 import io.getstream.chat.android.offline.experimental.errorhandler.factory.OfflineErrorHandlerFactory
 import io.getstream.chat.android.offline.experimental.plugin.configuration.Config
@@ -62,16 +60,6 @@ class ChatInitializer(private val context: Context) {
             .withPlugin(offlinePlugin)
             .withErrorHandler(offlineErrorHandlerFactory)
             .build()
-
-        InitializationCoordinator.getOrCreate().addUserConnectedListener {
-            val chatClient = ChatClient.instance()
-            ChatDomain.Builder(context, chatClient).apply {
-                enableBackgroundSync()
-                offlineEnabled()
-                userPresenceEnabled()
-                recoveryEnabled()
-            }.build()
-        }
 
         // Using markdown as text transformer
         ChatUI.messageTextTransformer = MarkdownTextTransformer(context)
